refactor: 拆分高级设置和导入Modal组件，增加JsonSchema导入校验
