+++ src/JsonSchemaEditor/SchemaItem/index.tsx
@@ -6,36 +6,27 @@ import {
   PlusOutlined,
   SettingOutlined,
 } from '@ant-design/icons';
+import AdvancedSettingModal from '@quiet-front-end/json-schema-editor-antd/JsonSchemaEditor/SchemaItem/AdvancedSettingModal';
+import ImportModal from '@quiet-front-end/json-schema-editor-antd/JsonSchemaEditor/SchemaItem/ImportModal';
 import {
   Button,
   Checkbox,
   Col,
   Dropdown,
-  Form,
   Input,
-  InputNumber,
   message,
-  Modal,
-  Radio,
   Row,
   Select,
-  Switch,
   theme,
   Tooltip,
 } from 'antd';
-import { Draft07 } from 'json-schema-library';
 import _ from 'lodash';
-import React, { useCallback, useEffect, useRef, useState } from 'react';
-import MonacoEditor from '../MonacoEditor';
+import React, { useCallback, useEffect, useState } from 'react';
 import { JSONSchema7 } from '../types';
 import {
   getDefaultSchema,
   getPropertyIndex,
-  parseJsonStr,
-  resolveJsonSchemaRef,
   SchemaTypeOptions,
-  SchemaTypes,
-  StringFormat,
 } from '../utils';
 
 type SchemaItemProps = {
@@ -68,7 +59,6 @@ type SchemaItemProps = {
 
 function SchemaItem(props: SchemaItemProps) {
   const { token } = theme.useToken();
-  const [advancedForm] = Form.useForm();
   const [messageApi, contextHolder] = message.useMessage();
   const {
     changeSchema,
@@ -83,7 +73,6 @@ function SchemaItem(props: SchemaItemProps) {
   } = props;
 
   const [schema, setSchema] = useState(props.schema);
-  const [formSchema, setFormSchema] = useState<any>();
   const [propertyName, setPropertyName] = useState(props.propertyName);
   const [schemaTitle, setSchemaTitle] = useState(schema.title);
   const [schemaDescription, setSchemaDescription] = useState(
@@ -98,15 +87,6 @@ function SchemaItem(props: SchemaItemProps) {
   const [expand, setExpand] = useState(true);
   const [advancedModal, setAdvancedModal] = useState(false);
   const [importModal, setImportModal] = useState(false);
-  const [importType, setImportType] = useState<'json' | 'json-schema'>('json');
-  const [importValue, setImportValue] = useState<string | undefined>();
-  const [isObject, setIsObject] = useState(false);
-  const [isArray, setIsArray] = useState(false);
-  const [isNumber, setIsNumber] = useState(false);
-  const [isBoolean, setIsBoolean] = useState(false);
-  const [isInteger, setIsInteger] = useState(false);
-  const [isString, setIsString] = useState(false);
-  const editorRef = useRef<any>(null);
   const isRoot = typeof propertyName === 'undefined';
 
   useEffect(() => {
@@ -141,23 +121,6 @@ function SchemaItem(props: SchemaItemProps) {
       handleDebounce.cancel();
     };
   }, [handleDebounce]);
-
-  useEffect(() => {
-    if (!advancedModal || !formSchema) {
-      return;
-    }
-    advancedForm.setFieldsValue(formSchema);
-    setIsObject(formSchema.type === 'object');
-    setIsArray(formSchema.type === 'array');
-    setIsNumber(formSchema.type === 'number');
-    setIsBoolean(formSchema.type === 'boolean');
-    setIsInteger(formSchema.type === 'integer');
-    setIsString(formSchema.type === 'string');
-  }, [advancedModal, formSchema]);
-
-  function handleEditorDidMount(editor: any) {
-    editorRef.current = editor;
-  }
 
   const schemaItems: any = schema.items;
   const addChildItems =
@@ -295,7 +258,6 @@ function SchemaItem(props: SchemaItemProps) {
                   ) {
                     return;
                   }
-                  setFormSchema(schema);
                   setAdvancedModal(!advancedModal);
                 }}
               />
@@ -423,456 +385,35 @@ function SchemaItem(props: SchemaItemProps) {
           handleAdvancedSettingClick={handleAdvancedSettingClick}
         />
       )}
-      <Modal
-        title="高级设置"
-        width={900}
+
+      <AdvancedSettingModal
+        schema={schema}
         open={advancedModal}
-        okText={'保存'}
-        cancelText={'取消'}
-        onOk={() => {
+        onOk={(newSchema) => {
           if (!changeSchema) {
             return;
           }
           if (isRoot || schema.type === 'object') {
-            changeSchema(namePath, { ...schema, ...formSchema });
-            setAdvancedModal(!advancedModal);
+            changeSchema(namePath, { ...newSchema });
+            setAdvancedModal(false);
             return;
           }
-          advancedForm
-            .validateFields()
-            .then((values) => {
-              changeSchema(
-                namePath,
-                { ...schema, ...formSchema, ...values },
-                propertyName,
-              );
-              setAdvancedModal(!advancedModal);
-            })
-            .catch((errorInfo) => {
-              console.log('Failed:', errorInfo);
-            });
+          changeSchema(namePath, { ...newSchema }, propertyName);
+          setAdvancedModal(false);
         }}
-        onCancel={() => setAdvancedModal(!advancedModal)}
-      >
-        <Form
-          form={advancedForm}
-          onValuesChange={(_, allValues) => {
-            if (editorRef.current) {
-              editorRef.current.setValue(
-                JSON.stringify({ ...formSchema, ...allValues }, null, 2),
-              );
-            }
-          }}
-        >
-          {!isObject && SchemaTypes.indexOf(formSchema?.type) !== -1 && (
-            <div
-              style={{
-                borderLeft: `3px solid ${token.colorPrimary}`,
-                fontSize: 16,
-                fontWeight: 399,
-                paddingLeft: 8,
-                marginBottom: 13,
-              }}
-            >
-              基本设置
-            </div>
-          )}
-          {(isString || isNumber || isInteger || isBoolean) && (
-            <Row
-              justify={'start'}
-              align={'middle'}
-              style={{ marginBottom: 13 }}
-            >
-              <Col span={4} style={{ textAlign: 'right' }}>
-                默认值：
-              </Col>
-              <Col span={8}>
-                <Form.Item noStyle name={'default'}>
-                  {isString && (
-                    <Input
-                      style={{ width: '100%' }}
-                      placeholder={'请输入默认值'}
-                    />
-                  )}
-                  {(isNumber || isInteger) && (
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入默认值'}
-                    />
-                  )}
-                  {isBoolean && (
-                    <Select
-                      style={{ width: '100%' }}
-                      placeholder={'请选择默认值'}
-                      options={[
-                        { value: true, label: 'true' },
-                        { value: false, label: 'false' },
-                      ]}
-                    />
-                  )}
-                </Form.Item>
-              </Col>
-            </Row>
-          )}
-          {isString && (
-            <Row
-              justify={'start'}
-              align={'middle'}
-              style={{ marginBottom: 13 }}
-            >
-              <Col span={4} style={{ textAlign: 'right' }}>
-                最小长度：
-              </Col>
-              <Col span={8}>
-                <Form.Item noStyle name={'minLength'}>
-                  <InputNumber
-                    min={0}
-                    style={{ width: '100%' }}
-                    parser={(value) =>
-                      value ? parseInt(value.replace(/\D/g, ''), 10) : ''
-                    }
-                    formatter={(value) =>
-                      value ? `${Math.floor(Math.max(value, 0))}` : ''
-                    }
-                    placeholder={'请输入最小长度'}
-                  />
-                </Form.Item>
-              </Col>
-              <Col span={4} style={{ textAlign: 'right' }}>
-                最大长度：
-              </Col>
-              <Col span={8}>
-                <Form.Item noStyle name={'maxLength'}>
-                  <InputNumber
-                    min={0}
-                    style={{ width: '100%' }}
-                    parser={(value) =>
-                      value ? parseInt(value.replace(/\D/g, ''), 10) : ''
-                    }
-                    formatter={(value) =>
-                      value ? `${Math.floor(Math.max(value, 0))}` : ''
-                    }
-                    placeholder={'请输入最大长度'}
-                  />
-                </Form.Item>
-              </Col>
-            </Row>
-          )}
-          {(isNumber || isInteger) && (
-            <>
-              <Row
-                justify={'start'}
-                align={'middle'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  最小值：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle name={'minimum'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入最小值'}
-                    />
-                  </Form.Item>
-                </Col>
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  最大值：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle name={'maximum'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入最大值'}
-                    />
-                  </Form.Item>
-                </Col>
-              </Row>
-              <Row
-                justify={'start'}
-                align={'middle'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  排他最小值：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle name={'exclusiveMinimum'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入排他最小值'}
-                    />
-                  </Form.Item>
-                </Col>
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  排他最大值：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle name={'exclusiveMaximum'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      placeholder={'请输入排他最大值'}
-                    />
-                  </Form.Item>
-                </Col>
-              </Row>
-            </>
-          )}
-          {isString && (
-            <>
-              <Row
-                justify={'start'}
-                align={'middle'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  正则匹配：
-                </Col>
-                <Col span={20}>
-                  <Form.Item noStyle name={'pattern'}>
-                    <Input placeholder={'请输入正则匹配公式'} />
-                  </Form.Item>
-                </Col>
-              </Row>
-              <Row
-                justify={'start'}
-                align={'middle'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  格式：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle name={'format'}>
-                    <Select
-                      allowClear
-                      options={StringFormat}
-                      placeholder={'请选择字符串格式'}
-                      style={{ width: '100%' }}
-                    />
-                  </Form.Item>
-                </Col>
-              </Row>
-            </>
-          )}
-          {isArray && (
-            <>
-              <Row
-                justify={'start'}
-                align={'middle'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  元素唯一：
-                </Col>
-                <Col span={20}>
-                  <Form.Item
-                    noStyle
-                    name={'uniqueItems'}
-                    valuePropName="checked"
-                  >
-                    <Switch />
-                  </Form.Item>
-                </Col>
-              </Row>
-              <Row
-                justify={'start'}
-                align={'middle'}
-                style={{ marginBottom: 13 }}
-              >
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  最少元素个数：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle name={'minItems'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      parser={(value) =>
-                        value ? parseInt(value.replace(/\D/g, ''), 10) : ''
-                      }
-                      formatter={(value) =>
-                        value ? `${Math.floor(Math.max(value, 0))}` : ''
-                      }
-                      placeholder={'请输入最少元素个数'}
-                    />
-                  </Form.Item>
-                </Col>
-                <Col span={4} style={{ textAlign: 'right' }}>
-                  最多元素个数：
-                </Col>
-                <Col span={8}>
-                  <Form.Item noStyle name={'maxItems'}>
-                    <InputNumber
-                      style={{ width: '100%' }}
-                      parser={(value) =>
-                        value ? parseInt(value.replace(/\D/g, ''), 10) : ''
-                      }
-                      formatter={(value) =>
-                        value ? `${Math.floor(Math.max(value, 0))}` : ''
-                      }
-                      placeholder={'请输入最多元素个数'}
-                    />
-                  </Form.Item>
-                </Col>
-              </Row>
-            </>
-          )}
-          {(isString || isNumber || isInteger) && (
-            <Row
-              justify={'start'}
-              align={'middle'}
-              style={{ marginBottom: 13 }}
-            >
-              <Col span={4} style={{ textAlign: 'right' }}>
-                枚举：
-              </Col>
-              <Col span={20}>
-                <Form.List name="enums">
-                  {(fields, { add, remove }) => (
-                    <>
-                      <Row gutter={6}>
-                        {fields.map(({ key, name, ...restField }) => (
-                          <Col span={12} key={key}>
-                            <Row
-                              gutter={2}
-                              justify={'start'}
-                              align={'middle'}
-                              style={{ marginBottom: 6 }}
-                            >
-                              <Col flex={'auto'}>
-                                <Form.Item
-                                  {...restField}
-                                  noStyle
-                                  name={[name]}
-                                  rules={[{ required: true }]}
-                                >
-                                  {isString && (
-                                    <Input placeholder="请输入枚举值" />
-                                  )}
-                                  {(isNumber || isInteger) && (
-                                    <InputNumber
-                                      style={{ width: '100%' }}
-                                      placeholder="请输入枚举值"
-                                    />
-                                  )}
-                                </Form.Item>
-                              </Col>
-                              <Col flex={'24px'}>
-                                <Button
-                                  icon={<DeleteOutlined />}
-                                  size={'small'}
-                                  type={'text'}
-                                  danger={true}
-                                  onClick={() => remove(name)}
-                                />
-                              </Col>
-                            </Row>
-                          </Col>
-                        ))}
-                      </Row>
-                      <Row>
-                        <Col span={6}>
-                          <Form.Item noStyle>
-                            <Button
-                              onClick={() => add()}
-                              block
-                              icon={<PlusOutlined />}
-                            >
-                              添加枚举值
-                            </Button>
-                          </Form.Item>
-                        </Col>
-                      </Row>
-                    </>
-                  )}
-                </Form.List>
-              </Col>
-            </Row>
-          )}
-          <div
-            style={{
-              borderLeft: `3px solid ${token.colorPrimary}`,
-              fontSize: 16,
-              fontWeight: 399,
-              paddingLeft: 8,
-              marginBottom: 13,
-            }}
-          >
-            Json Schema
-          </div>
-          <MonacoEditor
-            height={300}
-            language="json"
-            value={JSON.stringify(formSchema, null, 2)}
-            handleEditorDidMount={handleEditorDidMount}
-            onChange={(value) => {
-              handleDebounce(() => {
-                if (value) {
-                  try {
-                    const editorSchema = JSON.parse(value);
-                    setFormSchema(editorSchema);
-                  } catch (e) {}
-                }
-              });
-            }}
-          />
-        </Form>
-      </Modal>
+        onCancel={() => setAdvancedModal(false)}
+      />
 
-      <Modal
-        title="导入"
-        width={900}
-        okText={'导入'}
-        cancelText={'取消'}
+      <ImportModal
         open={importModal}
-        onOk={async () => {
-          if (!importValue || importValue.length === 0) {
-            messageApi.warning('请输入导入的 Json 数据');
-            return;
-          }
-          const importObject = parseJsonStr(importValue);
-          if (!importObject) {
-            messageApi.error('导入的内容不是 Json 格式的数据');
-            return;
-          }
-          let schema;
-          switch (importType) {
-            case 'json':
-              schema = new Draft07().createSchemaOf(importObject);
-              break;
-            case 'json-schema':
-              schema = await resolveJsonSchemaRef(importObject);
-              break;
-          }
+        onOk={(newSchema) => {
           if (changeSchema) {
-            changeSchema([], schema);
-            setImportModal(!importModal);
-            setImportValue(undefined);
+            changeSchema([], newSchema);
+            setImportModal(false);
           }
         }}
-        onCancel={() => setImportModal(!importModal)}
-      >
-        <Row style={{ marginBottom: 16 }}>
-          <Radio.Group
-            value={importType}
-            optionType="button"
-            buttonStyle="solid"
-            onChange={(type) => setImportType(type.target.value)}
-            options={[
-              { value: 'json', label: 'Json' },
-              { value: 'json-schema', label: 'JsonSchema' },
-            ]}
-          />
-        </Row>
-        <Row>
-          <MonacoEditor
-            height={390}
-            language="json"
-            handleEditorDidMount={handleEditorDidMount}
-            onChange={(value) => setImportValue(value)}
-          />
-        </Row>
-      </Modal>
+        onCancel={() => setImportModal(false)}
+      />
     </>
   );
 }
